+++ examples/eidolon_examples/security/authentication_processor.yaml
@@ -3,4 +3,4 @@ kind: Reference
 metadata:
   name: AuthenticationProcessor
 
-spec: GoogleJWTMiddleware
+spec: MSFTJWTProcessor
